Added swt 4.4 x64 library Added console only splash (UI support to come) Added in Main, commandline parser and -H option for Headless Jmule
Changed addTag behaviour, when a null tag (invalid) is generated it is
not added to the tag list (I have still to add logging). The previous
behaviour have failed with a NullPointerException. (I have to think
about deeply, perhaps the stricter parameter is better.. or not)

Added TODO and comments